[feat]  챌린지 완료 카드 공유하기 추가 및 extension으로 분리
